Implement function to determine firefox.
