+++ src/utils/browser.d.ts
@@ -1,0 +1,6 @@
+export {};
+
+declare global {
+    // eslint-disable-next-line @typescript-eslint/no-extraneous-class
+    class InstallTrigger {}
+}
