+++ src/utils/browser.ts
@@ -1,0 +1,4 @@
+// eslint-disable-next-line @typescript-eslint/triple-slash-reference
+/// <reference path="./browser.d.ts" />
+
+export const isFirefox = (): boolean => typeof InstallTrigger !== 'undefined';
